refactor(structured-logs): extract general log parser and tokenization
